fix: correct REQUIREMENT_WORK_PRODUCTS import
